src/pyiturr5etc/requirements.txt
astropy==7.0.1
intervaltree==3.1.0
ipython==8.31.0
ipython==9.0.2
matplotlib==3.10.3
natsort==8.4.0
numpy==2.2.6
pandas==2.2.3
pdfplumber==0.9.0
Pint==0.24.4
python_docx==1.1.2
termcolor==3.1.0
Unidecode==1.3.8
